feat(S11P21D109-284): 거래내역 구현

See merge request s11-blockchain-nft-sub1/S11P21D109!135

+++ backend/spring-server/src/main/java/com/semonemo/spring_server/domain/blockchain/service/BlockChainService.java
@@ -20,6 +20,6 @@ public interface BlockChainService {
 
     BigInteger convertToSmallestUnit(BigInteger amount);
 
-    BigInteger convertFromSmallestUnit(BigInteger amount);
+    Long convertFromSmallestUnit(BigInteger amount);
 }
